+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.0.6
+
+- fix: remove copy ref to contents context command, not that useful.
+- fix: add copy to current page and choose at top or bottom.
+
 ## v0.0.5
 
 - fix: optimize cut content action using moveBlock